Add "control undead" spell to game <!-- Write **BELOW** The Headers and **ABOVE** The comments else it may not be viewable. -->
## About The Pull Request

Adds a new spell to control zombies

Adds spell to necromancer's starting spells

## Why It's Good For The Game

zombies are currently mindless killers, and thus interaction with them for a necromancer is minimal and hostile.
This is good for the game, as it improves and expands on the concept of a zombie and perhaps the possibility of further RP for a necromancer, creating a zombie army and whatnot.

Currently, necromancers can only spawn limited and temporary skeletons which are unsuited for long-term defense and offense.

## Changelog

:cl:
add: Added 'control undead' spell
add: Added 'control undead' spell to necromancer's starting spells
code: Coded in 'control undead' spell
/:cl:

## Pre-Merge Checklist

- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [ ] You documented all of your changes.